requirements.txt
# CrewAI multi-agent framework
crewai

# LangChain modules for RAG
langchain
langchain-community
langchain-text-splitters

# Vector DB for embeddings
faiss-cpu

# PDF loading
pypdf

# Hugging Face embeddings
sentence-transformers

# OpenAI API client for Hugging Face Router
openai>=1.3.0

# Streamlit UI
streamlit

# Environment variable loader
python-dotenv

# YAML support
pyyaml

# Optional: data manipulation
numpy
